move transfer matrix + add argument
