+++ mttl/config.py
@@ -365,7 +365,6 @@ class TrainingArgs(DataArgs):
     seed: int = 42
     debug: bool = False
 
-    eval_before_training: bool = True
     precision: str = "32"
     monitor_grad_alignment_on: str = None
 
@@ -393,6 +392,7 @@ class TrainingArgs(DataArgs):
     eval_mmlu_flag: bool = False  # eval mmlu performance during training
     eval_rouge_flag: bool = False  # eval rouge during training
     eval_before_training: bool = True
+    create_transfer_matrix: bool = False
     pipeline_eval_tasks: str = None
     save_if_loaded_from_ckpt: bool = True
     dataset_type: str = None
@@ -510,7 +510,6 @@ class EvaluationConfig(MultiExpertConfig, TransformArgs):
     merge_or_route: str = None  # "uniform", "ties", "clown"
     tasksets_path: str = None
     remove_experts: str = None
-    create_transfer_matrix: bool = False
     es_metric: str = "loss"
     n_ng_iterations: int = 30  # number of iterations for LoraHub
     recompute_prototypes: bool = False
